src/pyproject.toml
[tool.poetry]
name = "roche"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
transitions = "^0.9.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
python-dotenv = "^1.0.0"
redis = "^5.0.1"
websockets = "^11.0.3"
mock = "^5.1.0"
pytest-mock = "^3.11.1"
fakeredis = "^2.19.0"
pytest-asyncio = "^0.21.1"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
pre-commit = "^3.3.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
flake8 = "^6.1.0"
black = "^23.9.1"
pytest = "^7.4.2"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
